refactor(frontend): move all UI JS/CSS under node package
