+++ frontend/src/legacy/js/components/media_card.js
