+++ frontend/src/legacy/js/components/media_type_toggle.js
