+++ frontend/src/legacy/js/views/calendar.js
