+++ frontend/src/legacy/js/views/dashboard.js
