+++ frontend/src/legacy/js/views/explore.js
