+++ frontend/src/legacy/js/views/inspector.js
