+++ frontend/src/legacy/js/views/item_detail.js
