+++ frontend/src/legacy/js/views/library.js
