+++ frontend/src/legacy/js/views/trending.js
