+++ frontend/src/legacy/js/views/vfs_stats.js
