+++ frontend/src/viewLoaders.ts
@@ -1,13 +1,13 @@
-import * as calendarView from "../../src/static/js/views/calendar.js";
-import * as dashboardView from "../../src/static/js/views/dashboard.js";
-import * as exploreView from "../../src/static/js/views/explore.js";
-import * as inspectorView from "../../src/static/js/views/inspector.js";
-import * as itemDetailView from "../../src/static/js/views/item_detail.js";
-import * as libraryView from "../../src/static/js/views/library.js";
-import * as mountView from "../../src/static/js/views/mount.js";
-import * as settingsView from "../../src/static/js/views/settings.js";
-import * as trendingView from "../../src/static/js/views/trending.js";
-import * as vfsStatsView from "../../src/static/js/views/vfs_stats.js";
+import * as calendarView from "./legacy/js/views/calendar.js";
+import * as dashboardView from "./legacy/js/views/dashboard.js";
+import * as exploreView from "./legacy/js/views/explore.js";
+import * as inspectorView from "./legacy/js/views/inspector.js";
+import * as itemDetailView from "./legacy/js/views/item_detail.js";
+import * as libraryView from "./legacy/js/views/library.js";
+import * as mountView from "./legacy/js/views/mount.js";
+import * as settingsView from "./legacy/js/views/settings.js";
+import * as trendingView from "./legacy/js/views/trending.js";
+import * as vfsStatsView from "./legacy/js/views/vfs_stats.js";
 import type { RouteName, ViewLoaderModule } from "./types";
 
 export const VIEW_LOADERS: Record<RouteName, ViewLoaderModule> = {
