+++ src/static/README.md
@@ -5,7 +5,7 @@
 - Source app: `/frontend`
 - Build command: `make frontend-build` (or `npm --prefix frontend run build`)
 - Output served by FastAPI: `src/static/ui/index.html` and `src/static/ui/assets/*`
-
-Legacy JS/CSS modules in `src/static/js` and `src/static/css` are still used by the React shell during migration.
+- Backend serves only the frontend bundle at `/` and `/static/ui/*`
 
 The previous Jinja2 template-based UI in `src/templates` has been removed.
+All frontend source JS/CSS is now managed under `frontend/src`.
